Moved syntax to a package object
